refactor(handy-core): move VAD (silero, smoothed) into handy-core
