+++ src-tauri/src/audio_toolkit/vad/mod.rs
@@ -1,32 +1,2 @@
-use anyhow::Result;
-
-pub enum VadFrame<'a> {
-    /// Speech – may aggregate several frames (prefill + current + hangover)
-    Speech(&'a [f32]),
-    /// Non-speech (silence, noise). Down-stream code can ignore it.
-    Noise,
-}
-
-impl<'a> VadFrame<'a> {
-    #[inline]
-    pub fn is_speech(&self) -> bool {
-        matches!(self, VadFrame::Speech(_))
-    }
-}
-
-pub trait VoiceActivityDetector: Send + Sync {
-    /// Primary streaming API: feed one 30-ms frame, get keep/drop decision.
-    fn push_frame<'a>(&'a mut self, frame: &'a [f32]) -> Result<VadFrame<'a>>;
-
-    fn is_voice(&mut self, frame: &[f32]) -> Result<bool> {
-        Ok(self.push_frame(frame)?.is_speech())
-    }
-
-    fn reset(&mut self) {}
-}
-
-mod silero;
-mod smoothed;
-
-pub use silero::SileroVad;
-pub use smoothed::SmoothedVad;
+//! Re-export shim. Real implementation lives in `handy_core::vad`.
+pub use handy_core::vad::*;
